requirements.txt
mysql-connector-python
pyside6
python-dotenv

